+++ sdks/typescript/react/src/components/index.ts
@@ -154,6 +154,8 @@ export type {
   ConferenceViewScreenShare,
   ConferenceViewSettingsDialogProps,
   ConferenceViewWhiteboard,
+  ThemePalette,
+  ThemeTexture,
   EndScreenProps,
   GuidedTourProps,
   JoiningScreenProps,